O2F-1041 [Feat] 검증alert 추가

+++ src/components/member/MemberInfoComponent.vue
@@ -39,9 +39,12 @@ const setDefault = async (addressId: number) => {
 }
 
 const deleteAdd = async (addressId: number) => {
-  deleteAddress(addressId)
-  alert('배송지가 삭제되었습니다.')
-  window.location.reload()
+  if (confirm("정말로 삭제하시겠습니까?")) {
+    deleteAddress(addressId)
+    alert('배송지가 삭제되었습니다.')
+    window.location.reload()
+    }
+ 
 }
 
 onMounted(async () => {
